% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{add_anomaly}
\alias{add_anomaly}
\title{Add an anomaly component to specification}
\usage{
add_anomaly(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The specification object with an anomaly component.
}
\description{
Generic method for identified anomaly component (such as outliers) of a model.
}
\details{
The method is used in state space estimation and simulation
models for adding an anomaly component to the specification, separate
from the regressor components.
}
