% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{add_ar}
\alias{add_ar}
\title{Add an autoregressive (AR) component to specification}
\usage{
add_ar(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The specification object with an AR component.
}
\description{
Generic method for the AR component  of a model.
}
\details{
The method is used in state space estimation and simulation
models for adding an AR component to the specification.
}
