% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{add_arma}
\alias{add_arma}
\title{Add an autoregressive moving average (ARMA) component to specification}
\usage{
add_arma(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The specification object with an ARMA component.
}
\description{
Generic method for the ARMA component  of a model.
}
\details{
The method is used in state space estimation and simulation
models for adding an ARMA component to the specification.
}
