% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{add_custom}
\alias{add_custom}
\title{Add a custom component to specification}
\usage{
add_custom(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The specification object with a custom component.
}
\description{
Generic method for a custom component of a model.
}
\details{
The method is used in state space estimation and simulation
models for adding a custom component to the specification.
}
