% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{add_distribution}
\alias{add_distribution}
\title{Add an error distribution component to specification}
\usage{
add_distribution(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The specification object with a distribution component.
}
\description{
Generic method for the distribution component of a model.
}
\details{
The method is used in state space estimation and simulation
models for adding a specific distribution component to the specification.
}
