% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{add_holiday}
\alias{add_holiday}
\title{Add a holiday component to specification}
\usage{
add_holiday(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The specification object with a holiday component.
}
\description{
Generic method for the holiday component of a model.
}
\details{
The method is used in state space estimation and simulation
models for adding a holiday component to the specification.
}
