% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{add_ma}
\alias{add_ma}
\title{Add a moving average (MA) component to specification}
\usage{
add_ma(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The specification object with a MA component.
}
\description{
Generic method for the MA component  of a model.
}
\details{
The method is used in state space estimation and simulation
models for adding a MA component to the specification.
}
