% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{add_polynomial}
\alias{add_polynomial}
\title{Add a polynomial component to specification}
\usage{
add_polynomial(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The specification object with a polynomial component.
}
\description{
Generic method for the polynomial component  of a model.
}
\details{
The method is used in state space estimation and simulation
models for adding a polyomial component to the specification.
}
