% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{add_regressor}
\alias{add_regressor}
\title{Add a regressor component to specification}
\usage{
add_regressor(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The specification object with a regressor component.
}
\description{
Generic method for the regressor component of a model.
}
\details{
The method is used in state space estimation and simulation
models for adding a regressor component to the specification.
}
