% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{add_seasonal}
\alias{add_seasonal}
\title{Add a seasonal component to specification}
\usage{
add_seasonal(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The specification object with a seasonal component.
}
\description{
Generic method for the seasonal component of a model.
}
\details{
The method is used in state space estimation and simulation
models for adding a seasonal component to the specification.
}
