% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{add_transform}
\alias{add_transform}
\title{Add a data transformation to specification}
\usage{
add_transform(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The specification object with a data transformation.
}
\description{
Generic method for a data transformation of a model.
}
\details{
The method is used in state space estimation and simulation
models for adding a data transformation to the specification.
}
