% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{estimate}
\alias{estimate}
\title{Model Estimation}
\usage{
estimate(object, ...)
}
\arguments{
\item{object}{an object of the model specification.}

\item{...}{additional parameters passed to the method.}
}
\value{
An object holding the results of the estimated model.
}
\description{
Generic method for estimating (fitting) a model.
}
