% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{estimate_ad}
\alias{estimate_ad}
\title{Estimation method using AD}
\usage{
estimate_ad(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\description{
Estimates a model using automatic differentiation.
}
