% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{halflife}
\alias{halflife}
\title{Half Life}
\usage{
halflife(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\description{
Calculates and returns the half-life of a model.
}
\details{
The half life is defined as the period it
takes a series  to reach half its long-term average values.
}
