% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pit}
\alias{pit}
\title{Probability Integral Transform (PIT)}
\usage{
pit(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The conditional probabilities.
}
\description{
Generic method for calculating the conditional probability integral
transform given the data and estimated density
}
\details{
The PIT is essentially the probabilities returned from the cumulative
distribution function (*p) given the data and estimated value of the mean,
conditional standard deviation and any other distributional parameters.
}
