% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{estimate}
\alias{estimate}
\title{Estimation Method}
\usage{
estimate(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\description{
Estimation Method
}
