\name{transparent_color}
\alias{transparent_color}
\title{
Function for color transparency.
}
\description{
Creates transparent colors for plotting.
}
\usage{
transparent_color(col, alpha = 100)
}
\arguments{
\item{col}{
A character string of the color to make transparent.
}
\item{alpha}{
Transparency percentage.
}
}
\value{
A hex string of the transparent color.
}
\author{
Alexios Galanos
}
