% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsaggregate}
\alias{tsaggregate}
\title{Time Series Aggregation}
\usage{
tsaggregate(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\description{
General aggregation method for both observed series as well 
as certain models such as homogeneous coefficients whose dynamics aggregate (see 
tsvets for an example).
}
