\name{tsautoselect}
\alias{tsautoselect}
\title{
Automatic Model Selection
}
\description{
Given a grid of model choices or values, estimates each one
and returns a list with the highest ranked based on a choice
of measures.
}
\usage{
tsautoselect(object, \ldots)
}
\arguments{
\item{object}{
Usually a default specification object which
will be overridden by values in the selection
grid.
}
\item{\ldots}{
Additional arguments
}
}
\author{
Alexios Galanos
}