% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsbacktest}
\alias{tsbacktest}
\title{Time Series Model Backtesting}
\usage{
tsbacktest(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\description{
Backtesting of a time series model.
}
