% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsbenchmark}
\alias{tsbenchmark}
\title{Time Series Model Benchmarking}
\usage{
tsbenchmark(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\description{
Time Series Model Benchmarking
}
