% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tscokurt}
\alias{tscokurt}
\title{Co-Kurtosis}
\usage{
tscokurt(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The co-kurtosis tensor (folded or unfolded).
}
\description{
Generic method for the co-kurtosis of a model.
}
\details{
The method calculates the conditional co-kurtosis of a model.
Applications of this can be found in the Independent Factor Conditional 
Density Models.
}
