% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsconvert}
\alias{tsconvert}
\title{Conversion method from one model to another}
\usage{
tsconvert(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\description{
Special purpose function for converting one model to another.
}
