% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsconvolve}
\alias{tsconvolve}
\title{Convolution of Distributions}
\usage{
tsconvolve(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The convolved density.
}
\description{
Generic method for convolution of conditional distribution.
}
\details{
The method is meant to apply the Fast Fourier Transform
to the characteristic function of a distribution. Applications of this can
be found in the Independent Factor Conditional Density Models.
}
