% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tscor}
\alias{tscor}
\title{Correlation matrix.}
\usage{
tscor(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\description{
Extracts the correlation matrix from a multivariate model or the
autocorrelation matrix of a univariate model.
}
