% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tscoskew}
\alias{tscoskew}
\title{Co-Skewness}
\usage{
tscoskew(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The co-skewness tensor (folded or unfolded).
}
\description{
Generic method for the co-skewness of a model.
}
\details{
The method calculates the conditional co-skewness of a model.
Applications of this can be found in the Independent Factor Conditional 
Density Models.
}
