% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tscov}
\alias{tscov}
\title{Covariance matrix.}
\usage{
tscov(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\description{
Extracts the covariance matrix from a multivariate model or the
autocovariance matrix of a univariate model.
}
