\name{tsdiagnose}
\alias{tsdiagnose}
\title{
Extract diagnostic model information.
}
\description{
Extracts the diagnostics from an estimated model.
}
\usage{
tsdiagnose(object, \ldots)
}
\arguments{
\item{object}{
An object of some S3 class.
}
\item{\ldots}{
Additional arguments
}
}
\author{
Alexios Galanos
}
