\name{tsequation}
\alias{tsequation}
\title{
Model Equation Extractor
}
\description{
Extract the equation of a model into either latex or some other format.
}
\usage{
tsequation(object, \ldots)
}
\arguments{
\item{object}{
An object of some S3 class.
}
\item{\ldots}{
Additional arguments
}
}
\author{
Alexios Galanos
}