% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmethods-package.R
\docType{package}
\name{tsmethods-package}
\alias{tsmethods}
\alias{tsmethods-package}
\title{tsmethods: Time Series Methods}
\description{
Common methods for use in a time series probabilistic framework.
}
\keyword{internal}
