% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsmetrics}
\alias{tsmetrics}
\title{Time Series Performance Metrics Method}
\usage{
tsmetrics(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\description{
Generates time series performance metrics.
}
