% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsmoments}
\alias{tsmoments}
\title{Analytic Forecast Moments}
\usage{
tsmoments(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\description{
here available, returns the analytic (or approximated) moments 
of the forecast (usually mean and variance).
}
