% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsprofile}
\alias{tsprofile}
\title{Profile a time series model}
\usage{
tsprofile(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\description{
Profiles a time series model by simulating from the estimated 
DGP and measuring the performance under different levels of system noise.
}
