% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsreport}
\alias{tsreport}
\title{Report Method}
\usage{
tsreport(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
An object holding a ready to parse report.
}
\description{
Generic method for generating reports in different output formats.
}
