\name{tsspec}
\alias{tsspec}
\title{
Extract specification object from estimation object
}
\description{
Extracts the specification object from an S3 object.
}
\usage{
tsspec(object, \dots)
}
\arguments{
\item{object}{
An object of some S3 class.
}
\item{\dots}{
Additional arguments required.
}
}
\author{
Alexios Galanos
}
