% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{unconditional}
\alias{unconditional}
\title{Unconditional Value}
\usage{
unconditional(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The unconditional value or values (e.g. mean and variance).
}
\description{
Generic method for extracting the unconditional value of a model.
}
