% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{ensemble_modelspec}
\alias{ensemble_modelspec}
\title{Ensemble specification}
\usage{
ensemble_modelspec(...)
}
\arguments{
\item{...}{objects of either all of class \dQuote{tsmodel.predict} or 
\dQuote{tsmodel.distribution} representing the probabilistic forecasts for 
the same horizon of optionally different models on the same series and with 
same number of draws. It is expected that the predictive distributions are 
based on joint simulated draws passed to the \code{innov} argument in 
the \code{predict} function of the supporting models. Instead of \ldots 
it is also possible to pass a list of the objects.}
}
\value{
An object of class \dQuote{ensemble.spec}.
}
\description{
Creates and validates an ensemble specification.
}
