\name{estimate_ad}
\alias{estimate_ad}
\title{
Estimation method using AD
}
\description{
Estimates a model using automatic differentiation.
}
\usage{
estimate_ad(object, \dots)
}
\arguments{
\item{object}{
An object of some appopriate S3 class.
}
\item{\dots}{
Additional arguments.
}
}
\author{
Alexios Galanos
}
