\name{tsaggregate}
\alias{tsaggregate}
\title{
Time Series Aggregation
}
\description{
General aggregation method for both observed series as well as certain
models such as homogeneous coefficients whose dynamics aggregate (see
tsvets for an example).
}
\usage{
tsaggregate(object, \ldots)
}
\arguments{
\item{object}{
An object of some S3 class.
}
\item{\ldots}{
Additional arguments
}
}
\author{
Alexios Galanos
}