\name{tsbacktest}
\alias{tsbacktest}
\title{
Time Series Model Backtesting
}
\description{
Backtesting of a time series model.
}
\usage{
tsbacktest(object, \dots)
}
\arguments{
\item{object}{
An object representing a model specification class.
}
\item{\dots}{
Additional arguments required.
}
}
\author{
Alexios Galanos
}
