\name{tsbenchmark}
\alias{tsbenchmark}
\title{
Time Series Model Benchmarking
}
\description{
Saves a model for benchmarking.
}
\usage{
tsbenchmark(object, \dots)
}
\arguments{
\item{object}{
An object of some appopriate S3 class.
}
\item{\dots}{
Additional arguments.
}
}
\author{
Alexios Galanos
}

