% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsbenchmark}
\alias{tsbenchmark}
\title{Time Series Model Benchmarking}
\usage{
tsbenchmark(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
An object holding the model benchmark which can be printed
and referenced.
}
\description{
Generic method for benchmarking models.
}
