% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tscalibrate}
\alias{tscalibrate}
\title{Prediction Calibration method}
\usage{
tscalibrate(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
An object with details of the calibrated target.
}
\description{
Generic method for calibrating a model, targeting specific objectives.
}
