\name{tsconvert}
\alias{tsconvert}
\title{
Conversion method from one model to another
}
\description{
Special purpose function for converting one model to another.
}
\usage{
tsconvert(object, \ldots)
}
\arguments{
\item{object}{
An object of some S3 class.
}
\item{\ldots}{
Additional arguments
}
}
\author{
Alexios Galanos
}
