\name{tsdecompose}
\alias{tsdecompose}
\title{
Time Series Model Decomposition
}
\description{
Decomposition of an estimated time series object.
}
\usage{
tsdecompose(object, \dots)
}
\arguments{
\item{object}{
An object of some S3 class generated generated by \command{estimate()} or \command{predict()}.
}
\item{\dots}{
Additional argument required.
}
}
\author{
Alexios Galanos
}
