% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsdecompose}
\alias{tsdecompose}
\title{Time Series Model Decomposition}
\usage{
tsdecompose(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
An object of the model decomposition.
}
\description{
Generic method for decomposition of an estimated time series object.
}
