% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsdiagnose}
\alias{tsdiagnose}
\title{Extract diagnostic model information}
\usage{
tsdiagnose(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The model diagnostics.
}
\description{
Generic method for extracting the diagnostics from an object.
}
