% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R, R/methods.R
\name{tsensemble.ensemble.spec}
\alias{tsensemble.ensemble.spec}
\alias{tsensemble}
\title{Ensembles predictions using their posterior predictive or simulated distribution}
\usage{
\method{tsensemble}{ensemble.spec}(object, weights = NULL, ...)

tsensemble(object, weights = NULL, ...)
}
\arguments{
\item{object}{currently only dispatches based on objects of 
class \dQuote{ensemble.spec} which have been validated for ensembling.}

\item{weights}{a numeric vector of length equal to the length of 
the ensembling predictions which represent the ensembling weights.}

\item{...}{additional parameters passed to the method.}
}
\value{
An object of class \dQuote{tsmodel.predict} or \dQuote{tsmodel.distribution} 
depending on the input class in \code{\link{ensemble_modelspec}}.

The ensembled distribution.
}
\description{
Generic method for ensembles of posterior predictive or simulated distributions 
spanning the same forecast horizon.
}
\details{
Returns the weighted distribution, under the assumption that the 
predictions were generated using a joint error distribution whose values were 
passed to the \code{innov} argument of the predict function used for each model.
}
