% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsequation}
\alias{tsequation}
\title{Model Equation Extractor}
\usage{
tsequation(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The model equation(s) in a specified format.
}
\description{
Generic method for extracting the equation of a model into either 
latex or some other format.
}
