% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsfilter}
\alias{tsfilter}
\title{Online Time Series Filter}
\usage{
tsfilter(object, y, newxreg = NULL, ...)
}
\arguments{
\item{object}{an object.}

\item{y}{new outcome data to filter on.}

\item{newxreg}{new optional regressors to filter on.}

\item{...}{additional parameters passed to the method.}
}
\value{
The filtered value(s).
}
\description{
Generic method for (online) time series filtering.
}
