% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R, R/methods.R
\name{tsgrowth.tsmodel.predict}
\alias{tsgrowth.tsmodel.predict}
\alias{tsgrowth}
\title{Growth Calculation}
\usage{
\method{tsgrowth}{tsmodel.predict}(object, d = 1, type = c("diff", "simple", "log"), ...)

tsgrowth(object, ...)
}
\arguments{
\item{object}{an object.}

\item{d}{the period back to look at for growth calculations.}

\item{type}{the type of growth calculation. \dQuote{diff} is simply the 
difference in values over n periods, \dQuote{simple} if the 
rate of change and \dQuote{log} the difference in logs.}

\item{...}{additional parameters passed to the method.}
}
\value{
an object of class \dQuote{tsmodel.predict} transformed to a growth distribution.

The growth distribution.
}
\description{
Generic method for calculating the growth distribution from an object.
}
