% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmethods-package.R
\docType{package}
\name{tsmethods-package}
\alias{tsmethods}
\alias{tsmethods-package}
\title{tsmethods: Time Series Methods}
\description{
Generic methods for use in a time series probabilistic framework, allowing for a common calling convention across packages. Additional methods for time series prediction ensembles and probabilistic plotting of predictions is included. A more detailed description is available at \url{https://www.nopredict.com/packages/tsmethods} which shows the currently implemented methods in the 'tsmodels' framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.nopredict.com/packages/tsmethods}
  \item \url{https://github.com/tsmodels/tsmethods}
}

}
\author{
\strong{Maintainer}: Alexios Galanos \email{alexios@4dscape.com} [copyright holder]

}
\keyword{internal}
