\name{tsmetrics}
\alias{tsmetrics}
\title{
Time Series Performance Metrics Method
}
\description{
Generates time series performance metrics.
}
\usage{
tsmetrics(object, \dots)
}
\arguments{
\item{object}{
An object of some appopriate S3 class.
}
\item{\dots}{
Additional arguments.
}
}
\author{
Alexios Galanos
}

