\name{tsmoments}
\alias{tsmoments}
\title{
Analytic Forecast Moments
}
\description{
Where available, returns the analytic (or approximated) moments of the
forecast (usually mean and variance).
}
\usage{
tsmoments(object, \ldots)
}
\arguments{
\item{object}{
An object of some S3 class.
}
\item{\ldots}{
Additional arguments
}
}
\author{
Alexios Galanos
}