% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsmoments}
\alias{tsmoments}
\title{Analytic Moments}
\usage{
tsmoments(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The moments of the class it was applied to (e.g. estimated or predicted object).
}
\description{
Generic method for generating the analytic (or approximated) moments 
of the model or forecast (usually mean and variance).
}
