\name{tsprofile}
\alias{tsprofile}
\title{
Profile a time series model
}
\description{
Profiles a time series model by simulating from the estimated
DGP and measuring the performance under different levels of 
system noise.
}
\usage{
tsprofile(object, \ldots)
}
\arguments{
\item{object}{
An object of some S3 class.
}
\item{\ldots}{
Additional arguments
}
}
\author{
Alexios Galanos
}