% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsprofile}
\alias{tsprofile}
\title{Profile a time series model}
\usage{
tsprofile(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
The profiled model with information such as root mean squared error per parameter.
}
\description{
Generic method for profiling a time series model by simulating from the estimated 
data generating process and measuring the performance under different levels of system noise.
}
