% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsspec}
\alias{tsspec}
\title{Extract specification object from estimation object}
\usage{
tsspec(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
A specification object extracted from an estimated or other
model which retains this information.
}
\description{
Generic method for extracting the specification object from an S3 object.
}
